+++ packages/lila-webpack-config/src/analyze.js
@@ -3,6 +3,7 @@ import WebpackBar from 'webpackbar';
 import SpeedMeasurePlugin from 'speed-measure-webpack-plugin';
 import base from './base';
 import { defaultEntry } from '../../../util/constants';
+import { defaultExt } from './defaults';
 
 const { join } = path;
 
@@ -21,8 +22,12 @@ export default ({ lila, webpack, entry, cmd, config, makeType }) => {
 
   baseConfig.plugins.push(new WebpackBar(), new SpeedMeasurePlugin());
 
+  const { ext = defaultExt } = config;
+
   return {
-    entry: `${srcPath}/${entry === defaultEntry ? '' : `${entry}/`}index.js`,
+    entry: `${srcPath}/${
+      entry === defaultEntry ? '' : `${entry}/`
+    }index.${ext}`,
     output: {
       path: analyzePath,
       filename: 'index.js',