+++ packages/lila-webpack-config/src/dev.js
@@ -1,6 +1,7 @@
 import path from 'path';
 import base from './base';
 import { defaultEntry } from '../../../util/constants';
+import { defaultExt } from './defaults';
 
 const { join } = path;
 
@@ -15,10 +16,12 @@ export default ({ lila, webpack, entry, cmd, config, makeType }) => {
 
   baseConfig.plugins.unshift(new HotModuleReplacementPlugin());
 
+  const { ext = defaultExt } = config;
+
   return {
     entry: [
       'webpack-hot-middleware/client?reload=true',
-      `${srcPath}/${entry === defaultEntry ? '' : `${entry}/`}index.js`,
+      `${srcPath}/${entry === defaultEntry ? '' : `${entry}/`}index.${ext}`,
     ],
     output: {
       path: devPath,